API/requirements.txt
Flask
python-dotenv
flask-cors
pymongo
requests